lingbin-temp-update: temp-commit-at-20240106
lingbin-temp-update: temp-commit-at-20240204

lingbin-temp-update: temp-commit-at-20240215

lingbin-temp-update: temp-commit-at-20240215-2

lingbin-temp-update: temp-commit-at-20240216

lingbin-temp-update: temp-commit-at-20240217

lingbin-temp-update: temp-commit-at-20240224

lingbin-temp-update: temp-commit-at-20240226

lingbin-temp-update: temp-commit-at-20240308

lingbin-temp-update: temp-commit-at-20240316

lingbin-temp-update: temp-commit-at-20240331

lingbin-temp-update: temp-commit-at-20240417

lingbin-temp-update: temp-commit-at-20240420

lingbin-temp-update: temp-commit-at-20240426

lingbin-temp-update: temp-commit-at-20240514

lingbin-temp-update: temp-commit-at-20240516

lingbin-temp-update: temp-commit-at-20240623

lingbin-temp-update: temp-commit-at-20240627

lingbin-temp: fix Driver.h

lingbin-temp: fix VectorHasher.cpp for 20240514

lingbin-temp: fix VectorHasher.cpp for 20240630

lingbin-temp: fix VectorHasher.cpp for 20240702

Fix PeriodicStatsReporter.h on MocOS(clang)

lingbin-temp: temp-commit-at-20240712

lingbin-temp: temp-commit-at-20240714: MmapArea

lingbin-temp: temp-commit-at-20240714: MemoryAllocator

lingbin-temp: temp-commit-at-20240714: MemoryAllocator

lingbin-temp: temp-commit-at-20240717: Driver

lingbin-temp: temp-commit-at-20240720: Task

lingbin-temp: temp-commit-at-20240806: exec-ExchangeClient

lingbin-temp: temp-commit-at-20240806: type-Filter

lingbin-temp: temp-commit-at-20240806: connector

lingbin-temp: temp-commit-at-20240810: connector

lingbin-temp: temp-commit-at-20240823: Vector

lingbin-temp: temp-commit-at-20240825: Connector

lingbin-temp: temp-commit-at-20240826: HashProbe-remove copy

lingbin-temp: temp-commit-at-20241119: SelectivityVector

lingbin-temp: temp-commit-at-20241219

lingbin-temp: temp-commit-at-20241220

lingbin-temp: temp-commit-at-20241222

lingbin-temp: temp-commit-at-20241223

lingbin-temp: temp-commit-at-20250102

lingbin-temp: temp-commit-at-20250108

lingbin-temp: temp-commit-at-20250114

lingbin-temp: temp-commit-at-20250119

lingbin-temp: temp-commit-at-20250126

lingbin-temp: temp-commit-at-20250202 velox/exec

lingbin-temp: temp-commit-at-20250206 velox/core

lingbin-temp: temp-commit-at-20250206 velox/expression

lingbin-temp: temp-commit-at-20250206 velox/expression

lingbin-temp: temp-commit-at-20250206 velox/parse

lingbin-temp: temp-commit-at-20250206 velox/connectos

lingbin-temp: temp-commit-at-20250308

lingbin-temp: temp-commit-at-20250320: dwio/common/XxxStream

lingbin-temp: temp-commit-at-20250320: dwio/common/caching

lingbin-temp: temp-commit-at-20250326: dwio/common/caching

lingbin-temp: temp-commit-at-20250327: velox/common/caching

lingbin-temp: temp-commit-at-20250404: velox/common/caching

lingbin-temp: temp-commit-at-20250411: velox/common/caching

Fix compile

lingbin-temp: temp-commit-at-20250428: StringView

lingbin-temp: temp-commit-at-20250430: FloatingPointUtil

lingbin-temp: temp-commit-at-20250430: DecimalUtil

lingbin-temp: temp-commit-at-20250509: Variant

lingbin-temp: temp-commit-at-20250512: Timestamp

lingbin-temp: temp-commit-at-20250512: velox/exec

lingbin-temp: temp-commit-at-20250612: velox/exec

lingbin-temp: temp-commit-at-20250612: velox/vector

lingbin-temp: temp-commit-at-20250617: velox/vector

lingbin-temp: temp-commit-at-20250709: velox/vector

lingbin-temp: temp-commit-at-20250715: velox/type/Variant

lingbin-temp: temp-commit-at-20250728: velox/type/Type.h

Fix compilation

Fix compilation: Type.cpp

lingbin-temp: temp-commit-at-20250805: velox/type/Type.h

lingbin-temp: temp-commit-at-20250806: velox/type/Variant.cpp

lingbin-temp: temp-commit-at-20250806: velox/type/Variant.cpp

lingbin-temp: temp-commit-at-20250827: connectors

lingbin-temp: temp-commit-at-20250827: caching

lingbin-temp: temp-commit-at-20250909: caching

lingbin-temp: temp-commit-at-20250911: caching

lingbin-temp: temp-commit-at-20250911: dwio/common

lingbin-temp: temp-commit-at-20250911: dwio/common & velox/connectors

lingbin-temp: temp-commit-at-20250911: connnectors/hive

lingbin-temp: temp-commit-at-20250911: connnectors/hive

lingbin-temp: temp-commit-at-20250911: common/caching

lingbin-temp: temp-commit-at-20250911: common/caching

lingbin-temp: temp-commit-at-20251003: common/caching

lingbin-temp: temp-commit-at-20251009: common/caching

lingbin-temp: temp-commit-at-20251010: common/caching

lingbin-temp: temp-commit-at-20251014: common/caching

lingbin-temp: temp-commit-at-20251014: dwio/common

fix compile

lingbin-temp: temp-commit-at-20251014: dwio/common

lingbin-temp: temp-commit-at-20251023: dwio/common

lingbin-temp: temp-commit-at-202511-17: common/Enums.h

lingbin-temp: temp-commit-at-20251117: common/config

lingbin-temp: temp-commit-at-20251117: common/memory

lingbin-temp: temp-commit-at-20251211: common/memory

lingbin-temp: temp-commit-at-20251212: exec/Exchange

lingbin-temp: temp-commit-at-20260105: exec/Exchange

lingbin-temp: temp-commit-at-20260105: exec/Exchange

lingbin-temp: temp-commit-at-20260106: exec/Exchange

lingbin-temp: temp-commit-at-20260106: exec/Exchange

lingbin-temp: temp-commit-at-20260115: exec/OutputBuffer

lingbin-temp: temp-commit-at-20260126: exec/Driver

lingbin-temp: temp-commit-at-20260202: exec/GroupId.h

+++ velox/common/memory/tests/StreamArenaTest.cpp
@@ -110,9 +110,9 @@ TEST_F(StreamArenaTest, newRange) {
         testData.expectedContiguousAllocationSize +
             testData.expectedNonContiguousAllocationSize,
         arena->size());
-    const int allocateBytes =
+    const uint64_t allocateBytes =
         AllocationTraits::pageBytes(mmapAllocator_->numAllocated());
-    const int contiguousBytes =
+    const uint64_t contiguousBytes =
         AllocationTraits::pageBytes(mmapAllocator_->numExternalMapped());
     ASSERT_EQ(
         testData.expectedNonContiguousAllocationSize,
